requirements.txt
black
pytest
pytest-testmon
pytest-watcher
loguru
python-dotenv
markdownify
requests
beautifulsoup4
